src/back-end/requirements_backend.txt
fastapi>=0.104.0
uvicorn>=0.23.2
pydantic>=2.4.2
python-dotenv>=1.0.0
langchain>=0.0.267
langchain-community>=0.0.10
qdrant-client>=1.6.0
openai>=1.0.0
sentence-transformers>=2.2.2
huggingface-hub>=0.16.4 
typing-extensions>=4.11.1